refactor: 컴포넌트 패키지 구조 변경
