+++ app/src/main/java/woowacourse/payments/ui/newcard/NewCardScreen.kt
@@ -23,7 +23,7 @@ import woowacourse.payments.domain.CardExpirationDate
 import woowacourse.payments.domain.CardHolderName
 import woowacourse.payments.domain.CardNumber
 import woowacourse.payments.domain.CardPassword
-import woowacourse.payments.ui.components.PaymentCard
+import woowacourse.payments.ui.common.components.PaymentCard
 import woowacourse.payments.ui.newcard.components.CardExpirationDateTextField
 import woowacourse.payments.ui.newcard.components.CardHolderNameTextField
 import woowacourse.payments.ui.newcard.components.CardNumberTextField
