+++ app/src/main/java/woowacourse/payments/ui/newcard/components/CardExpirationDateTextField.kt
@@ -9,7 +9,6 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import woowacourse.payments.R
 import woowacourse.payments.domain.CardExpirationDate
-import woowacourse.payments.ui.components.LimitedLengthOutlinedTextField
 import woowacourse.payments.ui.transformation.GroupedVisualTransformation
 import java.time.format.DateTimeFormatter
