+++ app/src/main/java/woowacourse/payments/ui/newcard/components/CardNumberTextField.kt
@@ -9,7 +9,6 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import woowacourse.payments.R
 import woowacourse.payments.domain.CardNumber
-import woowacourse.payments.ui.components.LimitedLengthOutlinedTextField
 import woowacourse.payments.ui.transformation.GroupedVisualTransformation
 
 @Composable
