+++ app/src/main/java/woowacourse/payments/ui/newcard/components/LimitedLengthOutlinedTextField.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.components
+package woowacourse.payments.ui.newcard.components
 
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.OutlinedTextField
